First refactoring run 3
